% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotnlp.r
\name{plotnlp}
\alias{plotnlp}
\title{plotnlp
Draw the barplot of negative log10 p-values for coregulated transcription factors}
\usage{
custom<-c("GATA4","GTF2B","JUN","SOX9","SP1","NCOA1","TRERF1","ZNF653","NFYA","PROX1","SOX2","PIAS1","TBX19","SOX8","EDF1","PNRC2","PITX1","NRIP1","CXXC1","CTNNB1","AR")

res<-tfcrcalc(custom)

plotnlp(res)
}
\arguments{
\item{df}{a dataframe of results obtained with tfcrcalc function}
}
\description{
plotnlp
Draw the barplot of negative log10 p-values for coregulated transcription factors
}
\examples{
custom<-c("GATA4","GTF2B","JUN","SOX9","SP1","NCOA1","TRERF1","ZNF653","NFYA","PROX1","SOX2","PIAS1","TBX19","SOX8","EDF1","PNRC2","PITX1","NRIP1","CXXC1","CTNNB1","AR")
res<-tfcrcalc(custom)
res<-head(res,n=15)
plotnlp(res)
}
