% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfcrnet.R
\name{tfcrnet}
\alias{tfcrnet}
\title{tfcrnet
built the network of tf coenriched}
\usage{
custom<-c("GATA4","GTF2B","JUN","SOX9","SP1","NCOA1","TRERF1","ZNF653","NFYA","PROX1","SOX2","PIAS1","TBX19","SOX8","EDF1","PNRC2","PITX1","NRIP1","CXXC1","CTNNB1","AR")

res<-tfcrcalc(custom)

tfcrnet(custom,res,layout=layout_as_star,cex=1,distance=2)
}
\arguments{
\item{custom}{a vector of human gene symbol containing possible tf}

\item{res}{a dataframe of results obtained with tfcrcalc function}

\item{layout}{string of layout type for the network (layout_as_star, layout_components, layout_in_circle, layout_nicely,layout_on_grid, layout_on_sphere, layout_randomly, layout_with_dh, layout_with_drl, layout_with_fr, layout_with_gem, layout_with_graphopt, layout_with_kk, layout_with_lgl, layout_with_mds)}

\item{cex}{size of the vertex labels}

\item{distance}{distance between vertex and its label}
}
\description{
tfcrnet
built the network of tf coenriched
}
\examples{
custom<-c("GATA4","GTF2B","JUN","SOX9","SP1","NCOA1","TRERF1","ZNF653","NFYA","PROX1","SOX2","PIAS1","TBX19","SOX8","EDF1","PNRC2","PITX1","NRIP1","CXXC1","CTNNB1","AR")
res<-tfcrcalc(custom)
res<-head(res,n=5)
res
tfcrnet(custom,res,layout=layout_nicely,cex=1,distance=1.5)
tfcrnet(custom,res,layout=layout_as_star,cex=1,distance=1.5)
}
